feat: Validação para impedir duplicação de autores e exclusão dos que possuem livros

+++ src/main/java/com/henriq/libraryapi/service/AuthorService.java
@@ -1,8 +1,11 @@
 package com.henriq.libraryapi.service;
 
 
+import com.henriq.libraryapi.exceptions.OperationNotAllowedException;
 import com.henriq.libraryapi.model.Author;
 import com.henriq.libraryapi.repository.AuthorRepository;
+import com.henriq.libraryapi.repository.BookRepository;
+import com.henriq.libraryapi.validator.AuthorValidator;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,12 +15,17 @@ import java.util.UUID;
 @Service
 public class AuthorService {
     private final AuthorRepository repository;
+    private final BookRepository bookRepository;
+    private final AuthorValidator validator;
 
-    public AuthorService(AuthorRepository repository){
+    public AuthorService(AuthorRepository repository, AuthorValidator validator, BookRepository bookRepository){
+        this.validator = validator;
+        this.bookRepository = bookRepository;
         this.repository = repository;
     }
 
     public Author save(Author author){
+        validator.validate(author);
         return this.repository.save(author);
     }
 
@@ -26,6 +34,9 @@ public class AuthorService {
     }
 
     public void delete(Author author){
+        if(hasABook(author)){
+            throw new OperationNotAllowedException("Autor(a) possui um livro cadastrado no sistema.");
+        }
         this.repository.delete(author);
     }
 
@@ -35,5 +46,9 @@ public class AuthorService {
         if(nationality != null) return this.repository.findByNationality(nationality);
 
         return repository.findAll();
+    }
+
+    public boolean hasABook(Author author){
+        return bookRepository.existsByAuthor(author);
     }
 }
